pptx-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.23
python-multipart==0.0.6
pillow==10.1.0
cairosvg==2.7.1
boto3==1.34.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1 